feat(Admin): adding react queries + update table + update lockedStatus
