+++ src/Components/AdminPanel/TableContent.tsx
@@ -1,63 +1,55 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 // libraries
-import { FC, useState, MouseEvent, useMemo, useCallback, Dispatch, SetStateAction } from 'react';
+import { FC, useState, MouseEvent, useMemo } from 'react';
 import { Checkbox, Table, TableBody, TableCell, TableContainer } from '@mui/material';
 // providers
 // files
 import TableHeader from './TableHeader';
 // import MockLanOrderList from '../../MockLanOrderList';
 import ascDescEnum from './ascDescEnum';
-import { StyledTableRow } from './AdminPanel';
+import { StyledTableRow } from './AdminPanelContent';
 import StyledTableCell from './StyledTableCell';
 import StyledTypeTableCell from './StyledTypeTableCell';
 import COLOURS from '../../Theme/Colours';
 import colourSwitch from './ColourSwitch';
-import axios from 'axios';
+// import axios from 'axios';
+import { basicOrderType } from '../../Context/Types';
 // styles
 
-const updateOrderCall = (userName: string, orderType: string, completed: boolean) => {
-  const newData = {
-    userName: userName,
-    orderType: orderType,
-    completed: completed,
-  };
-  try {
-    axios({
-      method: 'PUT',
-      url: `${import.meta.env.VITE_API_ADDRESS}Order`,
-      headers: {
-        'content-type': 'application/json',
-      },
-      data: newData,
-    }).catch((error) => console.log(error));
-  } catch (error) {
-    console.log(error);
-  }
-};
-
-export type Order = {
-  userName: string;
-  orderType: string;
-  completed: boolean;
-};
+// const updateOrderCall = (userName: string, orderType: string, completed: boolean) => {
+//   const newData = {
+//     userName: userName,
+//     orderType: orderType,
+//     completed: completed,
+//   };
+//   try {
+//     axios({
+//       method: 'PUT',
+//       url: `${import.meta.env.VITE_API_ADDRESS}Order`,
+//       headers: {
+//         'content-type': 'application/json',
+//       },
+//       data: newData,
+//     }).catch((error) => console.log(error));
+//   } catch (error) {
+//     console.log(error);
+//   }
+// };
 
 type TableContentProps = {
+  orderList: basicOrderType[];
   useColourMode: boolean;
   useColourModeWholeRow: boolean;
   strikethrough: boolean;
-  stateOrderList: Order[];
-  setStateOrderList: Dispatch<SetStateAction<Order[] | null>>;
 };
 
 const TableContent: FC<TableContentProps> = ({
+  orderList,
   useColourMode,
   useColourModeWholeRow,
   strikethrough,
-  stateOrderList,
-  setStateOrderList,
 }) => {
   const [orderDirection, setOrderDirection] = useState<ascDescEnum>(ascDescEnum.asc);
-  const [valueToOrderBy, setValueToOrderBy] = useState<keyof Order>('userName');
+  const [valueToOrderBy, setValueToOrderBy] = useState<keyof basicOrderType>('userName');
 
   function descendingComparator<T>(a: T, b: T, orderBy: keyof T) {
     if (b[orderBy] < a[orderBy]) {
@@ -79,7 +71,7 @@ const TableContent: FC<TableContentProps> = ({
       : (a, b) => -descendingComparator(a, b, orderBy);
   }
 
-  function stableSort<T>(array: Order[], comparator: (a: T, b: T) => number) {
+  function stableSort<T>(array: basicOrderType[], comparator: (a: T, b: T) => number) {
     const stabilizedThis = array.map((el, index) => [el, index] as [T, number]);
     stabilizedThis.sort((a, b) => {
       const order = comparator(a[0], b[0]);
@@ -91,30 +83,30 @@ const TableContent: FC<TableContentProps> = ({
     return stabilizedThis.map((el) => el[0]);
   }
 
-  const handleRequestSort = (_event: MouseEvent<unknown>, property: keyof Order) => {
+  const handleRequestSort = (_event: MouseEvent<unknown>, property: keyof basicOrderType) => {
     const isAsc = valueToOrderBy === property && orderDirection === ascDescEnum.asc;
     setOrderDirection(isAsc ? ascDescEnum.desc : ascDescEnum.asc);
     setValueToOrderBy(property);
   };
 
   const visibleRows = useMemo(
-    () => stableSort(stateOrderList, getComparator(orderDirection, valueToOrderBy)),
-    [getComparator, orderDirection, stateOrderList, valueToOrderBy],
+    () => stableSort(orderList, getComparator(orderDirection, valueToOrderBy)),
+    [getComparator, orderDirection, orderList, valueToOrderBy],
   );
 
-  const checkboxChangeHandler = useCallback(
-    (index: number, order: Order) => {
-      if (!stateOrderList && !order.completed) return;
+  // const checkboxChangeHandler = useCallback(
+  //   (index: number, order: basicOrderType) => {
+  //     if (!stateOrderList && !order.completed) return;
 
-      const completedFlip = !order.completed;
-      updateOrderCall(order.userName, order.orderType, completedFlip);
+  //     const completedFlip = !order.completed;
+  //     updateOrderCall(order.userName, order.orderType, completedFlip);
 
-      const updatedArray = [...stateOrderList];
-      updatedArray[index].completed = !updatedArray[index].completed;
-      setStateOrderList(updatedArray);
-    },
-    [setStateOrderList, stateOrderList],
-  );
+  //     const updatedArray = [...stateOrderList];
+  //     updatedArray[index].completed = !updatedArray[index].completed;
+  //     setStateOrderList(updatedArray);
+  //   },
+  //   [setStateOrderList, stateOrderList],
+  // );
 
   return (
     <TableContainer>
