+++ src/Context/Types.ts
@@ -4,3 +4,30 @@ export type Order = {
   orderType: string | null;
   completed: boolean | null;
 };
+
+export type basicOrderType = {
+  userName: string;
+  orderType: string;
+  completed: boolean;
+};
+
+export type PanelContainerProps = {
+  darkMode: boolean;
+  userName: string | undefined;
+  userDiscordId: string | undefined;
+  userLoggedIn: boolean;
+  lockedStatus: boolean;
+};
+
+export type PanelContentProps = {
+  darkMode: boolean;
+  userName: string;
+  userDiscordId: string;
+  lockedStatus: boolean;
+};
+
+export type BreakfastOption = {
+  name: string;
+  ingredients: string[];
+  colour: string;
+};
